refactor: extract shared PDF packet builder utility
Deduplicate PDF building logic between PatientEducationOutput and
PathwayLookup into a shared buildPatientPacket utility. Fix skeleton
loader to use deterministic widths instead of Math.random().
